+++ app/modules/application/index.php
@@ -2,10 +2,12 @@
 
 use GreenCheap\Application\Response;
 use GreenCheap\Application\UrlProvider;
-use GreenCheap\Kernel\ExceptionHandler;
-//use Symfony\Component\Debug\ErrorHandler;
 use Symfony\Component\ErrorHandler\Debug;
 use Symfony\Component\ErrorHandler\ErrorHandler;
+
+//@deprecated
+//use GreenCheap\Kernel\ExceptionHandler;
+//use Symfony\Component\Debug\ErrorHandler;
 
 return [
 
@@ -30,7 +32,8 @@ return [
         };
 
         //$app['exception'] = ExceptionHandler::register($app['debug']);
-        Debug::enable();
+
+        $app['exception'] = Debug::enable();
 
         ErrorHandler::register()->throwAt(E_ERROR | E_CORE_ERROR | E_COMPILE_ERROR | E_RECOVERABLE_ERROR);